+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- add missing checks in UserPreferences client
+
 ### Security
 
 ## [v1.53.0] - 2022-04-28